Allow only json requests when generating new URLs
Allows only .json requests to the home URL used for generating new
short URLs. This is important since there could potentially exist
a short URL named 'url', so it would collide with the home address
'localhost/url'. With this change requests to 'localhost/url.json'
can be safely routed to the controller that generates short URLs
while 'localhost/url' can be routed to the registered URL with the
code 'url'. This pattern will be used for all future URLs such as
'localhost/top' and 'localhost/.top.json'