Add json formatter to Url model
This ensures that json rendering of Url objects will only include relevant
attributes. This was made to minimize network data usage and as a security
measure to prevent internal system data to leak outside the system. For
example, data such as the internal database id of a record or its last
modified date should not be publicly available.