Add new Javascript front end for the API
I'll be adding a React as an interface to the API! The React app
will live along with the Rails API app to make deployment easier,
and also to be able to make calls between the front and back end
without worrying about CORS. The React app will, however, live
entirely inside the client folder, thus maintaining the separation
between the front end and back end apps. At the time of deplotment
the React app will be built into the public folder, so that Rails
can access it as an external resource. This way we keep
communication possible between both apps in a single package
without creating dependencies with one another. If at any time we
want to change the front end app all that needs to be done is to
delete the client folder.