Add method to parse URLs in backend
When the API must return a short URL it now always returns a full
URL with host and port, and not just the identifier. This makes
the backend and frontend more independent since previously the
frontend app had to build URLs. In the development environment
calls to the API use port 8081 by default, since there was a
bug when trying to read the current port. This means for now
rails must be run on port 8081 in the development environment.
This issue does not affect the production environment.