Reorganize project file structure
Three main categories were defined: components, which are generic
React components that can be reused anywhere within the app.
Scenes, which represent a single page in the app, along with any
specific components within that page. Lastly services, or reusable
non-React app logic, such as calls to the API. Each component and
scene is modular and includes everything it needs to work such as
any specific styles or media.